Moved some contents of the previous eventHandlers and helpers files into the index.js
